Updating document and contact components
