+++ .angular/cache/19.0.7/cms/vite/deps/@angular_common_http.js
@@ -1,0 +1,85 @@
+import {
+  FetchBackend,
+  HTTP_INTERCEPTORS,
+  HTTP_ROOT_INTERCEPTOR_FNS,
+  HTTP_TRANSFER_CACHE_ORIGIN_MAP,
+  HttpBackend,
+  HttpClient,
+  HttpClientJsonpModule,
+  HttpClientModule,
+  HttpClientXsrfModule,
+  HttpContext,
+  HttpContextToken,
+  HttpErrorResponse,
+  HttpEventType,
+  HttpFeatureKind,
+  HttpHandler,
+  HttpHeaderResponse,
+  HttpHeaders,
+  HttpInterceptorHandler,
+  HttpParams,
+  HttpRequest,
+  HttpResponse,
+  HttpResponseBase,
+  HttpStatusCode,
+  HttpUrlEncodingCodec,
+  HttpXhrBackend,
+  HttpXsrfTokenExtractor,
+  JsonpClientBackend,
+  JsonpInterceptor,
+  REQUESTS_CONTRIBUTE_TO_STABILITY,
+  provideHttpClient,
+  withFetch,
+  withHttpTransferCache,
+  withInterceptors,
+  withInterceptorsFromDi,
+  withJsonpSupport,
+  withNoXsrfProtection,
+  withRequestsMadeViaParent,
+  withXsrfConfiguration
+} from "./chunk-WRCFVEXO.js";
+import "./chunk-RB43LT5D.js";
+import "./chunk-HAYGR63U.js";
+import "./chunk-H4JBMRPK.js";
+export {
+  FetchBackend,
+  HTTP_INTERCEPTORS,
+  HTTP_TRANSFER_CACHE_ORIGIN_MAP,
+  HttpBackend,
+  HttpClient,
+  HttpClientJsonpModule,
+  HttpClientModule,
+  HttpClientXsrfModule,
+  HttpContext,
+  HttpContextToken,
+  HttpErrorResponse,
+  HttpEventType,
+  HttpFeatureKind,
+  HttpHandler,
+  HttpHeaderResponse,
+  HttpHeaders,
+  HttpParams,
+  HttpRequest,
+  HttpResponse,
+  HttpResponseBase,
+  HttpStatusCode,
+  HttpUrlEncodingCodec,
+  HttpXhrBackend,
+  HttpXsrfTokenExtractor,
+  JsonpClientBackend,
+  JsonpInterceptor,
+  provideHttpClient,
+  withFetch,
+  withInterceptors,
+  withInterceptorsFromDi,
+  withJsonpSupport,
+  withNoXsrfProtection,
+  withRequestsMadeViaParent,
+  withXsrfConfiguration,
+  HTTP_ROOT_INTERCEPTOR_FNS as ɵHTTP_ROOT_INTERCEPTOR_FNS,
+  HttpInterceptorHandler as ɵHttpInterceptingHandler,
+  HttpInterceptorHandler as ɵHttpInterceptorHandler,
+  REQUESTS_CONTRIBUTE_TO_STABILITY as ɵREQUESTS_CONTRIBUTE_TO_STABILITY,
+  withHttpTransferCache as ɵwithHttpTransferCache
+};
+//# sourceMappingURL=@angular_common_http.js.map
