+++ .angular/cache/19.0.7/cms/vite/deps/@angular_platform-browser.js
@@ -36,7 +36,8 @@ import {
   withI18nSupport,
   withIncrementalHydration,
   withNoHttpTransferCache
-} from "./chunk-SJJO32B2.js";
+} from "./chunk-4VELLTIB.js";
+import "./chunk-WRCFVEXO.js";
 import {
   getDOM
 } from "./chunk-RB43LT5D.js";
