+++ src/app/app.module.ts
@@ -22,6 +22,7 @@ import { AppRoutingModule } from './app-routing.module';
 import { DocumentEditComponent } from './documents/document-edit/document-edit.component';
 import { ContactEditComponent } from './contacts/contact-edit/contact-edit.component';
 import { ContactsFilterPipe } from './contacts/contacts-filter.pipe';
+import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
   declarations: [
@@ -49,6 +50,7 @@ import { ContactsFilterPipe } from './contacts/contacts-filter.pipe';
     FormsModule,
     ReactiveFormsModule,
     DragDropModule,
+    HttpClientModule,
   ],
 
   providers: [ContactService],
