+++ src/app/documents/document-detail/document-detail.component.html
@@ -18,7 +18,7 @@
       <br />
       <div class="row">
         <label>URL:</label>
-        <span class="label-value"> {{ document.imageUrl }}</span>
+        <span class="label-value"> {{ document.url }}</span>
       </div>
     </div>
   </div>
