+++ src/app/documents/document-edit/document-edit.component.html
@@ -44,7 +44,7 @@
             class="form-control"
             size="150"
             max="255"
-            [ngModel]="document?.imageUrl"
+            [ngModel]="document?.url"
             #imageUrl="ngModel"
             required
           />
